# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * IMUZAKI DONATION HUB API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  /** @nullable */
  message?: string | null;
}

export interface AdminLoginInput {
  username: string;
  password: string;
}

export interface AdminUser {
  id: number;
  username: string;
  email: string;
  createdAt?: string;
}

export interface AuthResponse {
  token: string;
  user: AdminUser;
}

export type DonationStatus = typeof DonationStatus[keyof typeof DonationStatus];


export const DonationStatus = {
  pending: 'pending',
  paid: 'paid',
  failed: 'failed',
  cancelled: 'cancelled',
  refunded: 'refunded',
} as const;

export interface Donation {
  id: number;
  donorName: string;
  isAnonymous?: boolean;
  amount: number;
  currency: string;
  /** @nullable */
  message?: string | null;
  status: DonationStatus;
  paymentMethod?: string;
  /** @nullable */
  transactionId?: string | null;
  createdAt: string;
  updatedAt?: string;
}

export interface DonationInput {
  /** @minLength 1 */
  donorName: string;
  isAnonymous?: boolean;
  /** @minimum 1 */
  amount: number;
  currency: string;
  message?: string;
}

export type DonationUpdateStatus = typeof DonationUpdateStatus[keyof typeof DonationUpdateStatus];


export const DonationUpdateStatus = {
  pending: 'pending',
  paid: 'paid',
  failed: 'failed',
  cancelled: 'cancelled',
  refunded: 'refunded',
} as const;

export interface DonationUpdate {
  status?: DonationUpdateStatus;
}

export interface DonationCreated {
  donation: Donation;
  paymentUrl: string;
}

export interface PaginatedDonations {
  data: Donation[];
  total: number;
  page: number;
  limit: number;
}

export interface TopDonor {
  donorName: string;
  totalAmount: number;
  donationCount: number;
  /** @nullable */
  lastDonationAt?: string | null;
}

export interface Goal {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  targetAmount: number;
  currentAmount: number;
  currency: string;
  isActive: boolean;
  /** @nullable */
  endsAt?: string | null;
  createdAt?: string;
}

export interface GoalInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  /** @minimum 1 */
  targetAmount: number;
  currency: string;
  isActive?: boolean;
  endsAt?: string;
}

export interface GoalUpdate {
  title?: string;
  description?: string;
  targetAmount?: number;
  currentAmount?: number;
  currency?: string;
  isActive?: boolean;
  /** @nullable */
  endsAt?: string | null;
}

export interface AnalyticsSummary {
  todayRevenue: number;
  weekRevenue: number;
  monthRevenue: number;
  totalRevenue: number;
  totalDonations: number;
  averageDonation: number;
  highestDonation: number;
}

export interface RevenueDataPoint {
  label: string;
  revenue: number;
  count: number;
}

export interface DashboardSummary {
  analytics: AnalyticsSummary;
  recentDonations: Donation[];
  topDonors: TopDonor[];
  activeGoal: Goal | null;
}

export interface Settings {
  id: number;
  streamerName: string;
  /** @nullable */
  streamerBio?: string | null;
  currency: string;
  minDonationAmount: number;
  alertDuration: number;
  theme: string;
  overlayAlertEnabled?: boolean;
  soundEnabled?: boolean;
  /** @nullable */
  activeSoundId?: number | null;
  updatedAt?: string;
}

export interface SettingsUpdate {
  streamerName?: string;
  streamerBio?: string;
  currency?: string;
  minDonationAmount?: number;
  alertDuration?: number;
  theme?: string;
  overlayAlertEnabled?: boolean;
  soundEnabled?: boolean;
  /** @nullable */
  activeSoundId?: number | null;
}

export interface Notification {
  id: number;
  type: string;
  title: string;
  message: string;
  isRead: boolean;
  /** @nullable */
  donationId?: number | null;
  createdAt: string;
}

export interface ActivityLog {
  id: number;
  action: string;
  description: string;
  /** @nullable */
  ipAddress?: string | null;
  /** @nullable */
  userId?: number | null;
  createdAt: string;
}

export interface PaginatedActivityLogs {
  data: ActivityLog[];
  total: number;
  page: number;
  limit: number;
}

export interface PaymentInput {
  donationId: number;
  amount: number;
  currency: string;
}

export interface PaymentResponse {
  paymentUrl: string;
  transactionId: string;
}

export interface PaymentCallbackInput {
  tran_id: string;
  status: string;
  /** @nullable */
  hash?: string | null;
  /** @nullable */
  amount?: number | null;
  /** @nullable */
  currency?: string | null;
}

export type PaymentStatusStatus = typeof PaymentStatusStatus[keyof typeof PaymentStatusStatus];


export const PaymentStatusStatus = {
  pending: 'pending',
  paid: 'paid',
  failed: 'failed',
  cancelled: 'cancelled',
  refunded: 'refunded',
} as const;

export interface PaymentStatus {
  donationId: number;
  status: PaymentStatusStatus;
  /** @nullable */
  transactionId?: string | null;
  /** @nullable */
  paidAt?: string | null;
}

export type OverlayDataSettings = {
  alertDuration: number;
  soundEnabled: boolean;
  streamerName: string;
};

export interface OverlayData {
  latestDonation: Donation | null;
  topDonation: Donation | null;
  activeGoal: Goal | null;
  settings: OverlayDataSettings;
}

export interface Sound {
  id: number;
  name: string;
  url: string;
  isActive: boolean;
  volume: number;
  createdAt?: string;
}

export interface SoundInput {
  /** @minLength 1 */
  name: string;
  url: string;
  isActive?: boolean;
  /**
     * @minimum 0
     * @maximum 100
     */
  volume?: number;
}

export interface SoundUpdate {
  name?: string;
  url?: string;
  isActive?: boolean;
  /**
     * @minimum 0
     * @maximum 100
     */
  volume?: number;
}

export type ListDonationsParams = {
page?: number;
limit?: number;
status?: ListDonationsStatus;
search?: string;
startDate?: string;
endDate?: string;
};

export type ListDonationsStatus = typeof ListDonationsStatus[keyof typeof ListDonationsStatus];


export const ListDonationsStatus = {
  pending: 'pending',
  paid: 'paid',
  failed: 'failed',
  cancelled: 'cancelled',
  refunded: 'refunded',
} as const;

export type GetRecentDonationsParams = {
limit?: number;
};

export type GetTopDonorsParams = {
limit?: number;
period?: GetTopDonorsPeriod;
};

export type GetTopDonorsPeriod = typeof GetTopDonorsPeriod[keyof typeof GetTopDonorsPeriod];


export const GetTopDonorsPeriod = {
  all: 'all',
  month: 'month',
  week: 'week',
} as const;

export type GetRevenueAnalyticsParams = {
period: GetRevenueAnalyticsPeriod;
};

export type GetRevenueAnalyticsPeriod = typeof GetRevenueAnalyticsPeriod[keyof typeof GetRevenueAnalyticsPeriod];


export const GetRevenueAnalyticsPeriod = {
  daily: 'daily',
  weekly: 'weekly',
  monthly: 'monthly',
  yearly: 'yearly',
} as const;

export type ListNotificationsParams = {
unreadOnly?: boolean;
};

export type ListActivityLogsParams = {
page?: number;
limit?: number;
};