# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: IMUZAKI DONATION HUB API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getAuthUser
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # DONATIONS
  /donations:
    get:
      operationId: listDonations
      tags: [donations]
      summary: List donations with filtering and pagination
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, paid, failed, cancelled, refunded]
        - name: search
          in: query
          schema:
            type: string
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated donations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedDonations"

    post:
      operationId: createDonation
      tags: [donations]
      summary: Create a new donation (public)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DonationInput"
      responses:
        "201":
          description: Donation created with payment URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DonationCreated"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /donations/recent:
    get:
      operationId: getRecentDonations
      tags: [donations]
      summary: Get recent donations (public)
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Recent donations list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Donation"
  /donations/top:
    get:
      operationId: getTopDonors
      tags: [donations]
      summary: Get top donors (public)
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: period
          in: query
          schema:
            type: string
            enum: [all, month, week]
            default: all
      responses:
        "200":
          description: Top donors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopDonor"
  /donations/{id}:
    get:
      operationId: getDonation
      tags: [donations]
      summary: Get donation by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Donation detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Donation"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateDonation
      tags: [donations]
      summary: Update donation status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DonationUpdate"
      responses:
        "200":
          description: Updated donation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Donation"

    delete:
      operationId: deleteDonation
      tags: [donations]
      summary: Delete donation (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # GOALS
  /goals:
    get:
      operationId: listGoals
      tags: [goals]
      summary: Get all donation goals (public)
      responses:
        "200":
          description: Goals list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Goal"

    post:
      operationId: createGoal
      tags: [goals]
      summary: Create a donation goal (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalInput"
      responses:
        "201":
          description: Created goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
  /goals/{id}:
    get:
      operationId: getGoal
      tags: [goals]
      summary: Get goal by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"

    patch:
      operationId: updateGoal
      tags: [goals]
      summary: Update goal (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalUpdate"
      responses:
        "200":
          description: Updated goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"

    delete:
      operationId: deleteGoal
      tags: [goals]
      summary: Delete goal (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ANALYTICS
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications (admin)
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    AdminLoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/AdminUser"
    AdminUser:
      type: object
      required: [id, username, email]
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        createdAt:
          type: string

    # DONATIONS
    Donation:
      type: object
      required: [id, donorName, amount, currency, status, createdAt]
      properties:
        id:
          type: integer
        donorName:
          type: string
        isAnonymous:
          type: boolean
        amount:
          type: number
        currency:
          type: string
        message:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, paid, failed, cancelled, refunded]
        paymentMethod:
          type: string
        transactionId:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    DonationInput:
      type: object
      required: [donorName, amount, currency]
      properties:
        donorName:
          type: string
          minLength: 1
        isAnonymous:
          type: boolean
        amount:
          type: number
          minimum: 1
        currency:
          type: string
        message:
          type: string
    DonationUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, paid, failed, cancelled, refunded]
    DonationCreated:
      type: object
      required: [donation, paymentUrl]
      properties:
        donation:
          $ref: "#/components/schemas/Donation"
        paymentUrl:
          type: string
    PaginatedDonations:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Donation"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    TopDonor:
      type: object
      required: [donorName, totalAmount, donationCount]
      properties:
        donorName:
          type: string
        totalAmount:
          type: number
        donationCount:
          type: integer
        lastDonationAt:
          type: ["string", "null"]

    # GOALS
    Goal:
      type: object
      required: [id, title, targetAmount, currentAmount, currency, isActive]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        targetAmount:
          type: number
        currentAmount:
          type: number
        currency:
          type: string
        isActive:
          type: boolean
        endsAt:
          type: ["string", "null"]
        createdAt:
          type: string
    GoalInput:
      type: object
      required: [title, targetAmount, currency]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        targetAmount:
          type: number
          minimum: 1
        currency:
          type: string
        isActive:
          type: boolean
        endsAt:
          type: string
    GoalUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        targetAmount:
          type: number
        currentAmount:
          type: number
        currency:
          type: string
        isActive:
          type: boolean
        endsAt:
          type: ["string", "null"]

    # ANALYTICS
...
    Notification:
      type: object
      required: [id, type, title, message, isRead, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
        title:
          type: string
        message:
          type: string
        isRead:
          type: boolean
        donationId:
          type: ["integer", "null"]
        createdAt:
          type: string

    # ACTIVITY LOGS
...