# lib/api-zod/src/generated/types/paginatedDonations.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * IMUZAKI DONATION HUB API
 * OpenAPI spec version: 0.1.0
 */
import type { Donation } from './donation';

export interface PaginatedDonations {
  data: Donation[];
  total: number;
  page: number;
  limit: number;
}